feat: new search/ls tools + improvements to existing tools (#11)
These changes should hopefully improve the efficiency of macher
operations - for example, it's now possible for the LLM to read parts of
files rather than only files in their entirety. The new search tool also
allows the LLM to be more dynamic in its investigation of the workspace.

More complete list of changes:

- Add a new grep-style search tool which uses `xref-matches-in-files` to
search by regexp

- Add a tool to list directory contents

- Add limit/offset/show-line-numbers parameters to the read tool

- Split the edit tool into separate edit/multi-edit tools, so the LLM
doesn't always have to pass an array of edits

- Improve path resolution during tool use, including more strict checks
to prevent the LLM from accessing files that aren't in the current
workspace's files list (even if they're under the workspace root, e.g.
gitignore'd files)

- Update tool/argument descriptions and the default context string that
gets passed with requests, to make the workspace abstraction (hopefully)
easier for the LLM to understand

- Internal/organizational cleanup for tool definitions

- More extensive tests for tool functionality