feat: more generic structure for macher action definitions (#16)
(this introduces some minor breaking API changes, see below)

Previously, macher actions were tied to a fairly specific workflow where
the prompt text gets pulled from the selected region or manual user
input, and then run through a transformer.

This change makes the action setup more generic: actions are now defined
as functions which can generate the full prompt however they like.
Action functions return a plist containing `:prompt` (the full prompt
text), `:preset` (the preset to use), and optionally `:summary`, which
can be used in the UI as a summarized version of the full prompt (for
built-in actions, the `:summary` is just the pre-transformed user
input).

The old behavior can be replicated using the
`macher-action-from-region-or-input` helper. For example, the built-in
definition for the `implement` action has been changed from this:

```emacs-lisp
'(implement :prompt "To implement: " :preset macher :transform macher--implement-prompt)
```

To this:

```emacs-lisp
`(implement
  .
  ,(apply-partially #'macher-action-from-region-or-input "To implement: " #'macher--implement-prompt))
```

`macher-action` now also supports passing an inline action configuration
instead of a symbol. This allows it to be used for "anonymous" actions
that aren't present in the global `macher-actions-alist`.

BREAKING CHANGES:

This change introduces some minor breaking API changes, which hopefully
don't affect too many users:

- If you've defined custom actions in the `macher-actions-alist`, they
need to be updated to use the new format. You can probably do something
similar to the above example of changes to the `implement` action. For
now, the old action format (a plist with `:prompt`, `:transform`, and
`:preset` keys) will be recognized and specially handled, but a warning
will be thrown and this fallback will be removed in a future update.

- The order of arguments to `macher-action` has been changed, to support
passing arbitrary arguments to action functions (this allows, for
example, input text to be provided programmatically when using the
default actions). This only affects you if you're calling
`macher-action` programmatically.

- The `input` field on the `macher-action-execution` struct has been
replaced with `summary`. This only affects you if you're working with
`macher-action-execution` objects programmatically, for example to
customize the UI using e.g. the `macher-before-action-functions`.

Interactive usage is otherwise unchanged.