feat: Flush cached context files
If _all_ changes to a workspace aren't done through `macher` (and if all
the changes proposed aren't immediately accepted) the context gets out
of sync with the buffer state. And then subsequent changes fail.

This change adds functions to flush the whole cache or just a particular
file. It also adds a custom setting `macher-auto-flush-cache` to do this
before actions and after context changes.